Force loading metadata before queries
