+++ frontend/src/metabase-types/store/mocks/dashboard.ts
@@ -15,9 +15,6 @@ export const createMockDashboardState = (
     startTime: null,
     endTime: null,
   },
-  loadingMetadata: {
-    loadingStatus: "idle",
-  },
   loadingControls: {},
   editingDashboard: null,
   isAddParameterPopoverOpen: false,
