+++ frontend/src/metabase/dashboard/actions/data-fetching.js
@@ -1,12 +1,7 @@
 import { getIn } from "icepick";
 import { t } from "ttag";
 
-import {
-  loadMetadataForDashcards,
-  loadMetadataForLinkedTargets,
-} from "metabase/dashboard/actions/metadata";
 import { showAutoApplyFiltersToast } from "metabase/dashboard/actions/parameters";
-import Dashboards from "metabase/entities/dashboards";
 import { defer } from "metabase/lib/promise";
 import { createAction, createThunkAction } from "metabase/lib/redux";
 import { equals } from "metabase/lib/utils";
@@ -42,9 +37,6 @@ export const FETCH_DASHBOARD_CARD_DATA =
   "metabase/dashboard/FETCH_DASHBOARD_CARD_DATA";
 export const CANCEL_FETCH_DASHBOARD_CARD_DATA =
   "metabase/dashboard/CANCEL_FETCH_DASHBOARD_CARD_DATA";
-
-export const FETCH_DASHBOARD_CARD_METADATA =
-  "metabase/dashboard/FETCH_DASHBOARD_CARD_METADATA";
 
 export const FETCH_CARD_DATA = "metabase/dashboard/FETCH_CARD_DATA";
 export const FETCH_CARD_DATA_PENDING =
@@ -378,25 +370,6 @@ export const fetchDashboardCardData =
       });
     }
   };
-
-export const fetchDashboardCardMetadata = createThunkAction(
-  FETCH_DASHBOARD_CARD_METADATA,
-  () => async (dispatch, getState) => {
-    const dashboard = getDashboardComplete(getState());
-    const dashboardType = getDashboardType(dashboard.id);
-    if (dashboardType === "normal") {
-      await dispatch(
-        Dashboards.actions.fetchMetadata({ id: dashboard.id }),
-      ).catch(error => {
-        console.error("Failed dashboard loading metadata", error);
-      });
-      await dispatch(loadMetadataForLinkedTargets(dashboard.dashcards));
-    }
-    if (dashboardType === "transient") {
-      await dispatch(loadMetadataForDashcards(dashboard.dashcards));
-    }
-  },
-);
 
 export const reloadDashboardCards = () => async (dispatch, getState) => {
   const dashboard = getDashboardComplete(getState());
