+++ frontend/src/metabase/dashboard/actions/revisions.js
@@ -1,7 +1,6 @@
 import {
   fetchDashboard,
   fetchDashboardCardData,
-  fetchDashboardCardMetadata,
 } from "metabase/dashboard/actions";
 import Revision from "metabase/entities/revisions";
 import { createThunkAction } from "metabase/lib/redux";
@@ -18,10 +17,9 @@ export const revertToRevision = createThunkAction(
           queryParams: null,
         }),
       );
-      await Promise.all([
-        dispatch(fetchDashboardCardMetadata()),
-        dispatch(fetchDashboardCardData({ reload: false, clearCache: true })),
-      ]);
+      await dispatch(
+        fetchDashboardCardData({ reload: false, clearCache: true }),
+      );
     };
   },
 );
