+++ frontend/src/metabase/dashboard/constants.ts
@@ -28,9 +28,6 @@ export const INITIAL_DASHBOARD_STATE: DashboardState = {
     startTime: null,
     endTime: null,
   },
-  loadingMetadata: {
-    loadingStatus: "idle",
-  },
   loadingControls: {},
   isAddParameterPopoverOpen: false,
   isNavigatingBackToDashboard: false,
