+++ frontend/src/metabase/dashboard/reducers.js
@@ -48,7 +48,6 @@ import {
   FETCH_CARD_DATA_PENDING,
   SET_DISPLAY_THEME,
   fetchDashboard,
-  FETCH_DASHBOARD_CARD_METADATA,
 } from "./actions";
 import { INITIAL_DASHBOARD_STATE } from "./constants";
 import {
@@ -465,30 +464,6 @@ const loadingDashCards = handleActions(
   INITIAL_DASHBOARD_STATE.loadingDashCards,
 );
 
-const loadingMetadata = handleActions(
-  {
-    [INITIALIZE]: {
-      next: state => ({
-        ...state,
-        loadingStatus: "idle",
-      }),
-    },
-    [FETCH_DASHBOARD_CARD_METADATA]: {
-      next: state => ({
-        ...state,
-        loadingStatus: "complete",
-      }),
-    },
-    [RESET]: {
-      next: state => ({
-        ...state,
-        loadingStatus: "idle",
-      }),
-    },
-  },
-  INITIAL_DASHBOARD_STATE.loadingMetadata,
-);
-
 const DEFAULT_SIDEBAR = { props: {} };
 const sidebar = handleActions(
   {
@@ -559,7 +534,6 @@ export const dashboardReducers = reduceReducers(
     parameterValues,
     draftParameterValues,
     loadingDashCards,
-    loadingMetadata,
     isAddParameterPopoverOpen,
     isNavigatingBackToDashboard,
     sidebar,
